Fix geoselection multiple requests
Remove geoOptions request from geoselection component and move it to its
parent pages: editEntryView/List.js and FilterEntriesForm.js
